add secretmemfd
This is a utility that I wrote many months ago; I didn't add it earlier
because memfd_secret did not work on my system. :/
After a recent kernel update, it started working.
I don't really have a use for memfd_secret in scripts now, especially
since those fds cannot be read, written, nor opened through procfs, they
can only be mmaped, and ftruncated; but I have been playing with
memfd_secret fds a bunch lately, now that it works on my kernel, and it
is conveninent to have a utility to create them.