add psendfd
This utility can replace a file descriptor of a given process with a
dup of a file descriptor of the current process.

It can also close a file descriptor of the target process if the FD
paramenter is -1. (Attempting to close a non-existing file descriptor
will cause an error, unless the -f option is specified)

Unfortunately, it does not work with pidfds because there is not a way
to ptrace a process with a pidfd. Also, annoyingly, it will not work if
the target process is being ptraced by another process like gdb or
strace.

Still, it is very cool!